Reconcile with the lower branches' fifth-round changes
The era fixture merges the stateless-stdio branch's split mode dispatch with
the mrtr-one-shot mode, which now counts as answered so the common handler
does not answer the same call again; the spec-side stdio transport accepts
the transport generation the stateless-stdio branch now passes to
send_request; the requests-only continuation pin reads the error this branch
raises; and the HTTP error-body helpers move into JsonRpcCommon::ErrorBodies
to keep the shared module within its length budget.
